+++ packages/next/src/client/components/router-reducer/reducers/server-action-reducer.ts
@@ -51,7 +51,7 @@ import { invalidateEntirePrefetchCache } from '../../segment-cache/cache'
 import { startRevalidationCooldown } from '../../segment-cache/scheduler'
 import { getDeploymentId } from '../../../../shared/lib/deployment-id'
 import { getNavigationBuildId } from '../../../navigation-build-id'
-import { NEXT_NAV_DEPLOYMENT_ID_HEADER } from '../../../../lib/constants-browser'
+import { NEXT_NAV_DEPLOYMENT_ID_HEADER } from '../../../../shared/lib/constants-browser'
 import {
   completeHardNavigation,
   convertServerPatchToFullTree,
