+++ packages/next/src/client/dev/hot-reloader/app/web-socket.ts
@@ -14,7 +14,7 @@ import {
 } from './hot-reloader-app'
 import { logQueue } from '../../../../next-devtools/userspace/app/forward-logs'
 import { InvariantError } from '../../../../shared/lib/invariant-error'
-import { WEB_SOCKET_MAX_RECONNECTIONS } from '../../../../lib/constants-browser'
+import { WEB_SOCKET_MAX_RECONNECTIONS } from '../../../../shared/lib/constants-browser'
 
 let reconnections = 0
 let reloading = false
