+++ packages/next/src/server/api-utils/node/api-resolver.ts
@@ -27,7 +27,7 @@ import {
   PRERENDER_REVALIDATE_HEADER,
   PRERENDER_REVALIDATE_ONLY_GENERATED_HEADER,
 } from '../../../lib/constants'
-import { JSON_CONTENT_TYPE_HEADER } from '../../../lib/constants-browser'
+import { JSON_CONTENT_TYPE_HEADER } from '../../../shared/lib/constants-browser'
 import { tryGetPreviewData } from './try-get-preview-data'
 import { parseBody } from './parse-body'
 import type { RevalidateFn } from '../../lib/router-utils/router-server-context'
