+++ packages/next/src/server/lib/patch-fetch.ts
@@ -5,7 +5,7 @@ import type {
 
 import { AppRenderSpan, NextNodeServerSpan } from './trace/constants'
 import { getTracer, SpanKind } from './trace/tracer'
-import { CACHE_ONE_YEAR_SECONDS } from '../../lib/constants-browser'
+import { CACHE_ONE_YEAR_SECONDS } from '../../shared/lib/constants-browser'
 import {
   INFINITE_CACHE,
   NEXT_CACHE_TAG_MAX_ITEMS,
