+++ packages/next/src/server/response-cache/index.test.ts
@@ -2,7 +2,7 @@ import ResponseCache from './index'
 import { CachedRouteKind, type ResponseCacheEntry } from './types'
 import { RouteKind } from '../route-kind'
 import RenderResult from '../render-result'
-import { HTML_CONTENT_TYPE_HEADER } from '../../lib/constants-browser'
+import { HTML_CONTENT_TYPE_HEADER } from '../../shared/lib/constants-browser'
 
 function mockIncrementalCache() {
   return {
