+++ packages/next/src/server/web/utils.ts
@@ -2,7 +2,7 @@ import type { OutgoingHttpHeaders } from 'http'
 import {
   NEXT_INTERCEPTION_MARKER_PREFIX,
   NEXT_QUERY_PARAM_PREFIX,
-} from '../../lib/constants-browser'
+} from '../../shared/lib/constants-browser'
 
 /**
  * Converts a Node.js IncomingHttpHeaders object to a Headers object. Any
