+++ packages/next/src/shared/lib/constants-browser.ts
