+++ packages/next/src/shared/lib/router/utils/route-regex.ts
@@ -1,7 +1,7 @@
 import {
   NEXT_INTERCEPTION_MARKER_PREFIX,
   NEXT_QUERY_PARAM_PREFIX,
-} from '../../../../lib/constants-browser'
+} from '../../../../shared/lib/constants-browser'
 import { INTERCEPTION_ROUTE_MARKERS } from './interception-routes'
 import { escapeStringRegexp } from '../../escape-regexp'
 import { removeTrailingSlash } from './remove-trailing-slash'
